we can now crossover with the point beyond the end of the first chromosome as long as it is not beyond the end of the second chromosome

+++ production/source/neat/SinglePointCrossover.h
@@ -11,7 +11,7 @@ private:
 public:
     explicit SinglePointCrossover();
 
-    Chromosome crossover(const Chromosome &a, const Chromosome &b, int crossoverPoint = 0);
+    Chromosome crossover(const Chromosome &a, const Chromosome &b, size_t crossoverPoint = 0);
 };
